requirements.txt
chromadb
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface
sentence-transformers
ollama
pypdf
textract
streamlit
streamlit-chat
pyyaml
tqdm
openai
